refactoring: routers -> 1)router_user 2)router_role
