+++ src/main.py
@@ -7,8 +7,8 @@ from src.auth.auth_config import auth_backend, current_user, fastapi_users
 from src.pages.routers import router as page_router
 from src.pages.routers import templates
 from src.users.models import User
-from src.users.routers import router as auth_router
-from src.users.routers import router_role
+from src.users.router_role import router_role
+from src.users.router_user import router as auth_router
 from src.users.schemas import UserCreate, UserRead
